Revert "refactor to reuse bulk writing usage for any update scenarios"
This reverts commit 691777fc2426ce9e46693d4fa35ecb7905790f5f.
